requirements.txt
openai
langchain
pdfplumber
python-dotenv
langchain_community 
dataparser